Setup to Study integration test in database
